+++ Source/LibationWinForms/ProcessQueue/ProcessQueueControl.resx
@@ -630,4 +630,7 @@
         w1EmAAAAAElFTkSuQmCC
 </value>
   </data>
+  <metadata name="counterTimer.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
+    <value>133, 17</value>
+  </metadata>
 </root>
